% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grtools.R
\name{CIplot}
\alias{CIplot}
\title{Plotting confidence intervals}
\usage{
CIplot(x, fit, ci, first = FALSE, xlab = "", ylab = "",
  xlim = NULL, ylim = NULL, col = 1, af = 0.3, border = NA,
  lwd = 1.5, lty = 1, density = 40, angle = 45, xaxt = "n",
  yaxt = "n", lwd.ci = 1, lty.ci = 1, ...)
}
\arguments{
\item{x, fit}{x and y argument of the fit}

\item{ci}{a matrix (\code{length(x)} x \code{2}) with upper and lower confidence intervals for the fit.}

\item{first}{a logical indicating if this is a first plot.}

\item{xlab, ylab, xlim, ylim, col, lwd, lty, xaxt, yaxt}{graphical parameters passed to
the \code{\link{plot}{graphics}} and the \code{\link{line}{graphics}} functions; see also \code{\link{par}{graphics}}.}

\item{af}{\code{alpha.f} of the \code{col} passed to the \code{\link{polygon}{graphics}}.}

\item{border, density, angle}{graphical parameters passed the \code{\link{polygon}{graphics}}.}

\item{lwd.ci, lty.ci}{lty and lwd eqivalents passed to the \code{\link{polygon}{graphics}}.}

\item{...}{other graphics parmeters.}
}
\description{
Plotting confidence intervals
}
\seealso{
fixPredict
}
