% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newpredict.R
\name{\%allin\%}
\alias{\%allin\%}
\title{All \%in\% function}
\usage{
x \%allin\% y
}
\arguments{
\item{x, y}{numeric vectors}
}
\description{
All \%in\% function
}
\author{
Maciej J. Danko
}
\keyword{internal}
