% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marPredict_e1.R
\docType{data}
\name{marPredict_e1}
\alias{marPredict_e1}
\title{Artificially generated data set}
\format{a data.frame with the fields
\describe{
\item{Y}{ response variable - counts. }
\item{linkY}{ internal: response variable on the link scale used to generate the data. }
\item{X1}{ a categorical variable. }
\item{X2}{ a categorical variable. }
\item{ID}{ a identification number of a subject. }
\item{IDv}{ internal: a random effect of a subject - its contribution to the intercept on the link scale. }
}}
\usage{
marPredict_e1
}
\description{
Artificially generated data set
}
\keyword{datasets}
