% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simglmer.R
\name{sim_glmer_data}
\alias{sim_glmer_data}
\title{Generating Poisson or binomial random intercept data (experimental function)}
\usage{
sim_glmer_data(formula = ~X1 * X2, theta = 0.75, coef = c(5, 0.5,
  0.7, -0.5, -0.7, 0.2, 0.3, 0.3, 0.2), coef.c = NULL,
  func.c = function(C1, coef.c) C1 * coef.c, s = 0,
  vcovmat = diag(rep(s * s, length(coef))), n.levels = c(3, 3),
  n.ID = 100, n.pop = 10000, family = c("Poisson", "binomial"), link)
}
\arguments{
\item{formula}{a formula used to generate the data; only fixed effect factors should be specified here.}

\item{theta}{a theta parameter for random intercept.}

\item{coef}{a vector of factor coefficients for resulting model matrix.}

\item{coef.c}{a value for coefficient for a continouse variable.}

\item{func.c}{a function that describes contribution of the continouse variable to the link predictor.
It takes two arguments: \code{C1} - a vector of values of continouse variable and \code{coef.c} - the parameter related to this variable.}

\item{s, vcovmat}{a residual standard deviation and a variance-covariance matrix for the fixed model coefficients.}

\item{n.levels}{a vector specifying a number of levels for each variable in the formula.}

\item{n.ID}{a number of levels (IDs) for random intercept.}

\item{n.pop}{a resulting sample size.}

\item{family}{a distribution family; either 'Poisson' or 'binomial'.}

\item{link}{a specification for the model link function. See \code{\link{family}{stats}}.}
}
\value{
a data frame with simulated data set
}
\description{
Generating Poisson or binomial random intercept data (experimental function)
}
\seealso{
fixPredict
}
